fix: Rename WearableUtility to AssetUtility
