+++ webapp/src/components/AssetPage/AssetUtility/AssetUtility.spec.tsx
@@ -1,9 +1,9 @@
 import { t } from 'decentraland-dapps/dist/modules/translation'
 import { renderWithProviders } from '../../../utils/test'
-import { WEARABLE_UTILITY_CONTENT_DATA_TEST_ID, WEARABLE_UTILITY_HEADER_DATA_TEST_ID, WearableUtility } from './WearableUtility'
-import { Props } from './WearableUtility.types'
+import { ASSET_UTILITY_CONTENT_DATA_TEST_ID, ASSET_UTILITY_HEADER_DATA_TEST_ID, AssetUtility } from './AssetUtility'
+import { Props } from './AssetUtility.types'
 
-const renderWearableUtility = (props: Partial<Props>) => renderWithProviders(<WearableUtility utility="someUtility" {...props} />)
+const renderWearableUtility = (props: Partial<Props>) => renderWithProviders(<AssetUtility utility="someUtility" {...props} />)
 
 describe('when rendering the wearable utility component', () => {
   let renderedComponent: ReturnType<typeof renderWearableUtility>
@@ -16,7 +16,7 @@ describe('when rendering the wearable utility component', () => {
 
   it('should render the utility header and the wearable utility', () => {
     const { getByTestId } = renderedComponent
-    expect(getByTestId(WEARABLE_UTILITY_HEADER_DATA_TEST_ID).textContent).toEqual(t('global.utility'))
-    expect(getByTestId(WEARABLE_UTILITY_CONTENT_DATA_TEST_ID).textContent).toEqual(utility)
+    expect(getByTestId(ASSET_UTILITY_HEADER_DATA_TEST_ID).textContent).toEqual(t('global.utility'))
+    expect(getByTestId(ASSET_UTILITY_CONTENT_DATA_TEST_ID).textContent).toEqual(utility)
   })
 })
