+++ webapp/src/components/AssetPage/AssetUtility/index.ts
@@ -1,0 +1,2 @@
+export * from './AssetUtility'
+export * from './AssetUtility.types'
