--- webapp/src/components/AssetPage/WearableUtility/index.ts
@@ -1,2 +1,0 @@
-export * from './WearableUtility'
-export * from './WearableUtility.types'
